further fixes
