+++ buildSrc/build.gradle.kts
@@ -1,7 +1,7 @@
 import org.gradle.kotlin.dsl.`kotlin-dsl`
 
 repositories {
-   jcenter()
+   mavenCentral()
 }
 
 plugins {
