+++ src/test/kotlin/io/kotest/assertions/arrow/core/NelMatchers.kt
@@ -1,38 +1,6 @@
 package io.kotest.assertions.arrow.core
 
 import arrow.core.NonEmptyList
-import io.kotest.assertions.arrow.nel.beSorted
-import io.kotest.assertions.arrow.nel.contain
-import io.kotest.assertions.arrow.nel.containAll
-import io.kotest.assertions.arrow.nel.containNoNulls
-import io.kotest.assertions.arrow.nel.containNull
-import io.kotest.assertions.arrow.nel.containOnlyNulls
-import io.kotest.assertions.arrow.nel.haveDuplicates
-import io.kotest.assertions.arrow.nel.haveElementAt
-import io.kotest.assertions.arrow.nel.haveSize
-import io.kotest.assertions.arrow.nel.shouldBeSingleElement
-import io.kotest.assertions.arrow.nel.shouldBeSorted
-import io.kotest.assertions.arrow.nel.shouldBeUnique
-import io.kotest.assertions.arrow.nel.shouldContain
-import io.kotest.assertions.arrow.nel.shouldContainAll
-import io.kotest.assertions.arrow.nel.shouldContainElementAt
-import io.kotest.assertions.arrow.nel.shouldContainNoNulls
-import io.kotest.assertions.arrow.nel.shouldContainNull
-import io.kotest.assertions.arrow.nel.shouldContainOnlyNulls
-import io.kotest.assertions.arrow.nel.shouldHaveDuplicates
-import io.kotest.assertions.arrow.nel.shouldHaveSize
-import io.kotest.assertions.arrow.nel.shouldNotBeSingleElement
-import io.kotest.assertions.arrow.nel.shouldNotBeSorted
-import io.kotest.assertions.arrow.nel.shouldNotBeUnique
-import io.kotest.assertions.arrow.nel.shouldNotContain
-import io.kotest.assertions.arrow.nel.shouldNotContainAll
-import io.kotest.assertions.arrow.nel.shouldNotContainElementAt
-import io.kotest.assertions.arrow.nel.shouldNotContainNoNulls
-import io.kotest.assertions.arrow.nel.shouldNotContainNull
-import io.kotest.assertions.arrow.nel.shouldNotContainOnlyNulls
-import io.kotest.assertions.arrow.nel.shouldNotHaveDuplicates
-import io.kotest.assertions.arrow.nel.shouldNotHaveSize
-import io.kotest.assertions.arrow.nel.singleElement
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.matchers.should
